refactor: move logLevel, formatter, memory
