--- packages/log/ts/createStandardLog.ts
@@ -1,5 +1,0 @@
-import { ConfigOptions } from './config.js'
-
-export function createStandardLog(options: Partial<ConfigOptions>) {
-  return options
-}
