+++ packages/log/ts/index.ts
@@ -3,11 +3,10 @@ export * from './configForTest.js'
 export * from './console/index.js'
 export * from './customLogLevel.js'
 export * from './errors.js'
-export * from './formatter/index.js'
+export * from './formatter.js'
 export * from './getLogger.js'
-export * from './logLevel.js'
-export * from './logLevelFn.js'
-export * from './memory/index.js'
+export * from './logLevels.js'
+export * from './memory.js'
 export * from './shouldLog.js'
 export * from './suppressLogs.js'
 export * from './types.js'
