--- packages/log/ts/logLevel.ts
@@ -1,57 +1,0 @@
-
-export const logLevels = {
-  /**
-   * none: 0
-   */
-  none: 0,
-  /**
-   * emergency: 100
-   */
-  emergency: 100,
-  /**
-   * alert: 200
-   */
-  alert: 200,
-  /**
-   * critical: 300
-   */
-  critical: 300,
-  /**
-   * error: 400
-   */
-  error: 400,
-  /**
-   * warn: 500
-   */
-  warn: 500,
-  /**
-   * notice: 600
-   */
-  notice: 600,
-  /**
-   * info: 700
-   */
-  info: 700,
-  /**
-   * debug: 800
-   */
-  debug: 800,
-  /**
-   * trace: 900
-   */
-  trace: 900,
-  /**
-   * planck: Infinity
-   * `planck unit` is a very small unit <https://en.wikipedia.org/wiki/Planck_units>.
-   */
-  planck: Infinity,
-  /**
-   * all: Infinity
-   */
-  all: Infinity
-}
-
-/**
- * @deprecated logLevel is renamed to logLevels
- */
-export const logLevel = logLevels
