+++ packages/log/ts/logLevels.ts
@@ -1,7 +1,58 @@
+export const logLevels = {
+  /**
+   * none: 0
+   */
+  none: 0,
+  /**
+   * emergency: 100
+   */
+  emergency: 100,
+  /**
+   * alert: 200
+   */
+  alert: 200,
+  /**
+   * critical: 300
+   */
+  critical: 300,
+  /**
+   * error: 400
+   */
+  error: 400,
+  /**
+   * warn: 500
+   */
+  warn: 500,
+  /**
+   * notice: 600
+   */
+  notice: 600,
+  /**
+   * info: 700
+   */
+  info: 700,
+  /**
+   * debug: 800
+   */
+  debug: 800,
+  /**
+   * trace: 900
+   */
+  trace: 900,
+  /**
+   * planck: Infinity
+   * `planck unit` is a very small unit <https://en.wikipedia.org/wiki/Planck_units>.
+   */
+  planck: Infinity,
+  /**
+   * all: Infinity
+   */
+  all: Infinity
+}
+
 import { reduceKey } from 'type-plus'
 import { getCustomLevel, getCustomLevelName, getCustomLevels } from './customLogLevel.js'
 import { getLoggers } from './getLoggers.js'
-import { logLevels } from './logLevel.js'
 import { store } from './store.js'
 import { Logger } from './types.js'
