+++ packages/log/ts/memory.ts
@@ -1,8 +1,8 @@
 import { required } from 'type-plus'
-import { formatLogLevel, plainLogFormatter } from '../formatter/index.js'
-import { toInspectLogEntry } from '../platform/index.js'
-import type { LogEntry, LogFilter, LogFormatter, LogReporter, LogReporterOptions } from '../types.js'
-import { assertLogModeIsNotProduction } from '../utils.js'
+import { formatLogLevel, plainLogFormatter } from './formatter.js'
+import { toInspectLogEntry } from './platform/index.js'
+import type { LogEntry, LogFilter, LogFormatter, LogReporter, LogReporterOptions } from './types.js'
+import { assertLogModeIsNotProduction } from './utils.js'
 
 export function logEntriesToString(logs: LogEntry[]) {
   return logs.map(toInspectLogEntry).map(plainLogFormatter).join('\n')
